pyproject.toml
[tool.poetry]
name = "ergo"
version = "0.1.0"
description = ""
authors = ["long2ice <long2ice@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
tortoise-orm = "^0.18.0"
starlette = { version = "^0.17.1", extras = ["full"] }
uvicorn = "^0.16.0"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
mypy = "*"
bandit = "*"
black = "*"
flake8 = "*"
isort = "*"
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py36', 'py37', 'py38', 'py39']
